+++ crates/adapters/celestia/src/test_helper/files.rs
@@ -582,7 +582,7 @@ fn filtered_block_from_path(
     let rollup_batch_shares = NamespaceRelevantData::new(batch_namespace, rollup_batch_rows);
     let rollup_proof_shares = NamespaceRelevantData::new(proof_namespace, rollup_proof_rows);
 
-    FilteredCelestiaBlock::new(rollup_batch_shares, rollup_proof_shares, header)
+    FilteredCelestiaBlock::new(rollup_batch_shares, rollup_proof_shares, header.into())
 }
 
 pub(crate) fn make_test_path(data_path: &str) -> PathBuf {