+++ crates/full-node/sov-sequencer/src/lib.rs
@@ -67,6 +67,9 @@ pub enum SequencerNotReadyDetails {
     },
     /// The sequencer is shutting down and cannot accept transactions.
     Shutdown,
+
+    /// The replica is waiting for the first batch from master.
+    ReplicaNotReady,
 }
 /// An object-safe interface to the sequencer, which can be used to
 /// publish a proof blob to DA.