+++ crates/full-node/sov-sequencer/src/preferred/mod.rs
@@ -39,6 +39,7 @@ use sov_modules_api::{
     ApiTxEffect, FullyBakedTx, RejectReason, Runtime, RuntimeEventProcessor, RuntimeEventResponse,
     Spec, StateCheckpoint, StateUpdateInfo, VersionReader, VisibleSlotNumber, *,
 };
+use sov_modules_stf_blueprint::PreExecError;
 use sov_rest_utils::errors::internal_server_error_500;
 use sov_rest_utils::errors::{database_error_500, sequencer_overloaded_503};
 use sov_rollup_interface::common::SlotNumber;
@@ -63,8 +64,8 @@ use transaction_subscriptions::TransactionCache;
 
 use crate::common::{
     error_not_fully_synced, generic_accept_tx_error, loop_send_tx_notifications, poll_state_update,
-    AcceptedTx, Sequencer, SequencerEventStream, StateUpdateError, StateUpdateNotification,
-    SubscriptionStreamError, WithCachedTxHashes,
+    pre_exec_err_to_accept_tx_err, AcceptedTx, Sequencer, SequencerEventStream, StateUpdateError,
+    StateUpdateNotification, SubscriptionStreamError, WithCachedTxHashes,
 };
 use crate::metrics::{track_in_progress_batch_size, PreferredSequencerFetchBatchesToReplayMetrics};
 use crate::preferred::block_executor::{RollupBlockExecutor, RollupBlockExecutorError};
@@ -237,6 +238,9 @@ where
             handles.push(worker);
         }
 
+        let (mut replica_task, start_replica_task_notifier) =
+            ReplicaSyncTask::new(shutdown_sender.clone()).await?;
+
         let tx_queue_id = Arc::new(AtomicU64::new(0));
         let (synchronized_state, synchronized_state_updator) = create(
             api_ledger_db.clone(),
@@ -252,7 +256,8 @@ where
             stop_at_rollup_height,
             rollup_exec_config.clone(),
             cached_txs.write_handle(),
-            cache_warm_up_executor.clone(),
+            cache_warm_up_executor,
+            start_replica_task_notifier,
         );
 
         let synchronized_state_task = synchronized_state.start().await;
@@ -290,17 +295,17 @@ where
             if let Some(postgres_connection_string) =
                 &config.sequencer_kind_config.postgres_connection_string
             {
-                let mut replica_task = ReplicaSyncTask::new(
-                    postgres_connection_string.clone(),
-                    shutdown_sender.clone(),
-                )
-                .await?;
-
-                let replica_task_handle = replica_task.start(synchronized_state_updator).await;
+                let replica_task_handle = replica_task
+                    .start(
+                        synchronized_state_updator,
+                        postgres_connection_string.clone(),
+                    )
+                    .await;
                 handles.push(replica_task_handle.data_fetcher_handle);
                 handles.push(replica_task_handle.sync_task_handle);
             }
         }
+
         handles.push(tokio::spawn({
             update_state_task(
                 seq.clone(),
@@ -857,18 +862,12 @@ where
 
         // Check if this transaction has a configured delay
         let runtime = Rt::default();
-        let call = match Rt::Auth::decode_serialized_tx(&baked_tx) {
-            Ok((call, _)) => call,
-            Err(_) => {
-                return Err(ErrorObject {
-                    status: StatusCode::BAD_REQUEST,
-                    message: "Unable to decode transaction".to_string(),
-                    details: sov_rest_utils::json_obj!({
-                        "error": "Unable to decode transaction".to_string(),
-                    }),
-                });
-            }
-        };
+        let mut state = self
+            .api_state()
+            .default_api_state_accessor()
+            .to_provable_reader();
+        let (_, _, call) = <Rt as Runtime<S>>::Auth::authenticate(&baked_tx, &mut state)
+            .map_err(|e| pre_exec_err_to_accept_tx_err(PreExecError::AuthError(e)))?;
         let call = Rt::wrap_call(call);
         let delay_ms = runtime.get_transaction_delay_ms(&call);