+++ examples/demo-rollup/tests/evm/evm_test_helper.rs
@@ -28,6 +28,7 @@ pub(crate) const SENDER_PRIV_KEY: &str =
 
 pub(crate) const EVM_EXTENSION: SeqConfigExtension = SeqConfigExtension {
     max_log_limit: 20000,
+    response_size_limit: (1024 * 1024) - (1024 * 30), // Limit our response size to 1MB, leaving 30kb for headers, overhead, and misestimation.
 };
 
 /// Starts test rollup node.