+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # 2025-11-01
+- #2013 Ensure `eth_getLogsWithCursor` respects a 1MB response size limit. Fix its cursor deserialization behavior to match other chains.
 - #2006 Add associated Error type to the EVM module.
 - #1953 Performance improvement for processing stf changes, but might delay writing finalized data to disk by 1 block
+- #2008 Treat `Latest` as `Pending` in EVM module to avoid foundry issues.
 
 # 2025-10-31
 - #2002 **Resync breaking change**. This PR moves rollup configuration files to demo-rollup/configs directory. 
+- #2007 Changes celestia rollup metrics, please switch to new dashboard.
 
 # 2025-10-29
 - #1996 Set gas limit to 1B on ETH API access.
@@ -22,6 +25,7 @@
 - #1966 **Resync breaking change**: Updates the schema generated for the `signature` and `public_key` fields on `Transaction` variants (both V0 and V1) to be byte arrays of the correct length. This is a follow-up on #1877, affecting only the web3 SDK functionality. However, this is a breaking change to the `CHAIN_HASH`, and therefore will invalidate existing transactions.
 
 # 2025-10-22
+- #1950 Adds a new new config option storage.separate_archival_state. If set to true, archival state will go into its own db at {storage_path}/archival-state-db. This option is false by default for backwards compatibility. When not set, the old database layout is used.
 - #1951 *Minor breaking change* Adds a new memo field to `Bank::TokenTransfer` events and a new callmessage which can set that field. This change requires resyncing any existing nodes due to the change in event serialization.
 - #1937 Reject log subscriptions with block filters.
 - #1887 *Minor breaking change*: Adds an optional third generic to the `Transaction` type, allowing overriding the `CryptoSpec` that defines the public key and signature types in the transaction. The generic defaults to `Spec::CryptoSpec`, which was the previous behaviour. This is a non-breaking change for the majority of cases, however some usages of `Transaction` may require explicitly specifying the `Runtime` and `Spec` generics which the compiler was able to infer previously.